refactor: use nx-workspace ang ng-packagr
