+++ libs/store/public_api.ts
@@ -1,0 +1,1 @@
+export * from './src/index';
