+++ libs/store/spec/decorators/pluck.spec.ts
@@ -1,11 +1,11 @@
-import { Select } from '@ngrx-utils/store';
-import { Observable } from 'rxjs/Observable';
-import { NgrxSelect } from '@ngrx-utils/store/src/decorators/ngrx-select.module';
+import { NgrxSelectModule, Pluck, Select } from '@ngrx-utils/store';
+import { NgrxSelect } from '../../src/decorators/ngrx-select.module';
 import { createFeatureSelector, createSelector, Store } from '@ngrx/store';
+import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
 import { map } from 'rxjs/operators';
 
-describe('Select', () => {
+describe('Pluck', () => {
   interface FooState {
     foo: boolean | null;
     bar?: {
@@ -36,43 +36,32 @@ describe('Select', () => {
   const msBar = createSelector(msFeature, state => state.bar);
 
   const store = new Store(of(globalState), undefined as any, undefined as any);
-
-  it('selects sub state with Select decorator', () => {
+  it('should select sub state with Pluck decorator', () => {
     NgrxSelect.store = store;
     class MyStateSelector {
-      @Select(msBar) bar$: Observable<any>; // using MemoizedSelector
+      @Pluck('myFeature', 'bar', 'a', 'b', 'c', 'd')
+      hello$: Observable<string>;
+      @Pluck() myFeature: Observable<FooState>; // implied by name
+      @Pluck('myFeature.bar.a.b.c.d') hi$: Observable<string>;
     }
 
     const mss = new MyStateSelector();
 
-    mss.bar$.subscribe(n => {
-      expect(n).toBe(globalState.myFeature.bar);
+    mss.hello$.subscribe(n => {
+      expect(n).toBe('world');
+    });
+
+    mss.myFeature.subscribe(n => {
+      expect(n).toBe(globalState.myFeature);
+    });
+    mss.hi$.subscribe(n => {
+      expect(n).toBe('world');
     });
   });
-
-  it('should apply pipeable operator when provided', () => {
-    NgrxSelect.store = store;
-    class MyStateSelector {
-      @Select(
-        msBar,
-        map(a => {
-          return !!a;
-        })
-      )
-      bar$: Observable<any>; // using MemoizedSelector
-    }
-
-    const mss = new MyStateSelector();
-
-    mss.bar$.subscribe(n => {
-      expect(n).toBe(true);
-    });
-  });
-
   it('should throw error when user not import NgrxSelectModule', () => {
     NgrxSelect.store = undefined;
     class MyStateSelector {
-      @Select(msBar) bar$: Observable<any>; // using MemoizedSelector
+      @Pluck('bar$') bar$: Observable<any>; // using MemoizedSelector
     }
 
     const mss = new MyStateSelector();
@@ -85,21 +74,7 @@ describe('Select', () => {
 
     expect(() => {
       class MyStateSelector {
-        @Select('msBar' as any)
-        bar$: Observable<any>; // using MemoizedSelector
-      }
-    }).toThrowError();
-  });
-
-  it('should throw error when property is non-configurable', () => {
-    function seal(target: any, name: string) {
-      Object.seal(target);
-    }
-    NgrxSelect.store = store;
-    expect(() => {
-      class MyStateSelector {
-        @Select(msBar)
-        @seal
+        @Pluck((() => {}) as any)
         bar$: Observable<any>; // using MemoizedSelector
       }
     }).toThrowError();
