+++ libs/store/spec/operators/pluck.spec.ts
