+++ libs/store/src/decorators/pluck.ts
