+++ libs/store/src/directives/index.ts
@@ -1,0 +1,2 @@
+export { NgLetDirective, NgLetModule } from './ngLet';
+export { RouterLinkActiveMatchModule, RouterLinkActiveMatch } from './routerLinkActiveMatch';
