+++ libs/store/src/index.ts
@@ -1,3 +1,3 @@
 export * from './decorators';
-export * from './operators';
 export * from './directives';
+export * from './operators';
