+++ libs/store/src/operators/index.ts
@@ -1,0 +1,2 @@
+export { pluck } from './pluck';
+export { untilDestroy } from './untilDestroy';
