+++ libs/store/src/operators/pluck.ts
