+++ libs/store/tsconfig.json
@@ -1,21 +1,17 @@
 {
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "baseUrl": ".",
     "declaration": true,
     "stripInternal": true,
-    "experimentalDecorators": true,
     "module": "es2015",
     "moduleResolution": "node",
     "outDir": "../../dist/store",
-    "rootDir": "./",
     "sourceMap": true,
     "inlineSources": true,
-    "target": "es2015",
-    "lib": ["es2015", "dom"],
-    "skipLibCheck": true,
-    "strict": true
+    "target": "es2015"
   },
   "files": ["public_api.ts"],
+  "include": ["**/*.ts"],
   "angularCompilerOptions": {
     "annotateForClosureCompiler": true,
     "strictMetadataEmit": true,
