--- modules/store/decorators/public_api.ts
@@ -1,1 +1,0 @@
-export * from '../src/decorators';
