--- modules/store/operators/public_api.ts
@@ -1,2 +1,0 @@
-export * from '../src/operators';
-
