--- modules/store/src/directives/index.ts
@@ -1,2 +1,0 @@
-export { NgLetDirective, NgLetModule } from './ngLet';
-export { RouterLinkActiveMatchModule, RouterLinkActiveMatch } from './routerLinkActiveMatch';
