--- modules/store/src/operators/index.ts
@@ -1,2 +1,0 @@
-export { pluck } from './pluck';
-export { untilDestroy } from './untilDestroy';
